+++ Sources/Tools/MacOS/TestMacOSTool.swift
@@ -63,6 +63,7 @@ public struct TestMacOSTool: Sendable {
                             ),
                         ]),
                     ].merging([String: Value].testSchemaProperties) { _, new in new }
+                        .merging([String: Value].withoutBuildingSchemaProperty) { _, new in new }
                         .merging([String: Value].continueBuildingSchemaProperty) { _, new in new }
                         .merging([String: Value].buildSettingsSchemaProperty) { _, new in new }
                         .merging([String: Value].extraArgsSchemaProperty) { _, new in new },
@@ -114,6 +115,7 @@ public struct TestMacOSTool: Sendable {
             scheme: scheme,
             destination: destination,
             configuration: configuration,
+            withoutBuilding: arguments.getBool("without_building"),
             additionalArguments: arguments.continueBuildingArgs()
                 + arguments.buildSettingOverrides() + extraArgs,
             environment: environment,