+++ Sources/Tools/Simulator/TestSimTool.swift
@@ -58,6 +58,7 @@ public struct TestSimTool: Sendable {
                             ),
                         ]),
                     ].merging([String: Value].testSchemaProperties) { _, new in new }
+                        .merging([String: Value].withoutBuildingSchemaProperty) { _, new in new }
                         .merging([String: Value].continueBuildingSchemaProperty) { _, new in new }
                         .merging([String: Value].buildSettingsSchemaProperty) { _, new in new }
                         .merging([String: Value].extraArgsSchemaProperty) { _, new in new },
@@ -101,6 +102,7 @@ public struct TestSimTool: Sendable {
             scheme: scheme,
             destination: destination,
             configuration: configuration,
+            withoutBuilding: arguments.getBool("without_building"),
             additionalArguments: arguments.continueBuildingArgs()
                 + arguments.buildSettingOverrides() + extraArgs,
             environment: environment,